# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Apollo14 Air Quality API",
        "description": "API for air quality monitoring and photo submissions",
        "contact": {
            "email": "support@apollo14.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/submissions/{id}": {
            "get": {
                "tags": [
                    "Submissions"
                ],
                "summary": "Get specific submission details",
                "description": "Retrieves detailed information about a specific submission",
                "operationId": "cd2a59547a628d683933a63fb3b892cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Submission ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 123
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 123
                                                },
                                                "user_guess": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "actual_level": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "accuracy_score": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "photo_url": {
                                                    "type": "string",
                                                    "example": "http://localhost:8000/storage/submissions/photo.jpg"
                                                },
                                                "location_lat": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 40.7128
                                                },
                                                "location_lng": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": -74.006
                                                },
                                                "air_quality_data": {
                                                    "type": "object"
                                                },
                                                "submitted_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-01-15T10:30:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Authentication required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Submission not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Submission not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}